+++ src/main/java/io/github/epi155/recfm/td/GrpTraitDescription.java
@@ -3,6 +3,7 @@ package io.github.epi155.recfm.td;
 import io.github.epi155.recfm.api.CodeFactory;
 import io.github.epi155.recfm.api.GrpTraitModel;
 import io.github.epi155.recfm.api.TraitModel;
+import io.github.epi155.recfm.proxy.GrpProxy;
 import org.yaml.snakeyaml.TypeDescription;
 import org.yaml.snakeyaml.nodes.Node;
 
@@ -16,8 +17,20 @@ public class GrpTraitDescription extends TypeDescription {
         substituteProperty("len", int.class, null, "setLength");
         substituteProperty("ovr", boolean.class, null, "setOverride");
         substituteProperty("as", TraitModel.class, null, "setTypedef");
+        substituteProperty("ref", String.class, null, "setReference");
     }
     public Object newInstance(Node node) {
-        return factory.newGrpTraitModel();
+        return new GrpProxy(factory);
     }
+
+    /* nella 2.2 si perde il return, ma è stato FIXato nella 2.5 */
+    @Override
+    public Object finalizeConstruction(Object obj) {
+        if (obj instanceof GrpProxy) {
+            GrpProxy proxy = (GrpProxy) obj;
+            return  proxy.getDelegate();
+        }
+        return obj;
+    }
+
 }