+++ src/main/java/io/github/epi155/recfm/td/OccTraitDescription.java
@@ -3,6 +3,7 @@ package io.github.epi155.recfm.td;
 import io.github.epi155.recfm.api.CodeFactory;
 import io.github.epi155.recfm.api.OccTraitModel;
 import io.github.epi155.recfm.api.TraitModel;
+import io.github.epi155.recfm.proxy.OccProxy;
 import org.yaml.snakeyaml.TypeDescription;
 import org.yaml.snakeyaml.nodes.Node;
 
@@ -17,8 +18,20 @@ public class OccTraitDescription extends TypeDescription {
         substituteProperty("ovr", boolean.class, null, "setOverride");
         substituteProperty("as", TraitModel.class, null, "setTypedef");
         substituteProperty("x", int.class, null, "setTimes");
+        substituteProperty("ref", String.class, null, "setReference");
     }
     public Object newInstance(Node node) {
-        return factory.newOccTraitModel();
+        return new OccProxy(factory);
     }
+
+    /* nella 2.2 si perde il return, ma è stato FIXato nella 2.5 */
+    @Override
+    public Object finalizeConstruction(Object obj) {
+        if (obj instanceof OccProxy) {
+            OccProxy proxy = (OccProxy) obj;
+            return  proxy.getDelegate();
+        }
+        return obj;
+    }
+
 }